// src/app/models/Feriados.ts
export class Feriados {
        feriadoId: number;
        sedeId: number;
        data: Date;
        descricao: string;
        usuarioIdInclusao: number;
        dataInclusao: Date;
        usuarioIdAlteracao: number;
        dataAlteracao: Date;
        usuarioIdExclusao: number;
        dataexclusao: Date;
        ativo: boolean;
        nomeSede: string;
}